requirements.txt
pyyaml==5.4.1
ruamel.yaml==0.17.4
pytest>=6.2.3
pylint>=2.8.1
